:recycle: Separate situp animation from state
